Sanitize repository
Visual Studio code settings were being commited because they contained a
list of ignored words for the spell checker and a task file.

In general, commiting IDE settings is a very bad idea, but in this case
it was even worse because those settings were related to an extension of
the IDE (cSpell), not even the IDE itself, and a building task.

The most unobstrusive way of dealing with spell-ignored words, at least
in this particular case, is to add them as comments with directives for
the spell checker (cSpell in this case).

Also, linter errors have been fixed and comments adjusted to fix a bunch
of spelling errors.