Enable wait for keypress by default
Waiting for a keypress when exiting is now enabled for all scripts when
running under Win32. This has been done because of two reasons. First is
that it can be disabled at will by any script not wanting the facility,
so having it enabled by default is not a problem.

Second (and most important), is that it is unobtrusive when the script
is run from the command line, since it won't do anything, but prevents
missing on-screen messages, and any other information, if the script is
run on a transient console, for example by double-clicking it, no matter
if intentionally or accidentally.

Affected scripts have been fixed.