Improve API documentation in README.md
The actual documentation comes from the code itself, but is is kept in
sync by hand, which for now is more than enough. In the future, maybe a
dedicated tool is needed for keep the documentation in sync, and until
that happens, the authoritative documentation is the one in the code and
the one in README.md is just a best effort.

Caveat emptor.