Improve output streams encoding problem fix
The encoding of both sys.stdout and sys.stderr was being fixed so it was
UTF-8 even when the streams are redirected to a file. Unfortunately, the
fix will fail catastrophically if the streams are overridden or replaced
by third party code or libraries imported before legion.

So now, before applying the encoding fix using reconfigure(), a check is
made to make sure the streams belong to the TextIOWrapper class.