Modify run() function signature
Until now, it could receive any number of strings as the command to run,
and optional keyword options for subprocess.run(). This led to excessive
complexity in type hinting, ambiguous usage and an ambiguous signature.

The reason behind that was to mimic subprocess.run() signature, but it
makes no sense. It is better to have a clearer, more explicit signature.

So, from now on, it expects a sequence of strings for the command to be
run, and the optional keyword options for subprocess.run().