Remove namespace for exportable constants
Until now, exportable constants were stored on its own namespace, to wit
a class named 'Constants'. Regarding code organization, this was a more
or less useful strategy.

Unfortunately, it has a few drawbacks:
- Since this is a module, the constants already had a namespace, which
is the module itself. Adding a second namespace is not needed.
- The code using this module was more verbose than necessary due to the
second namespace.
- All constants were imported, no matter if used or not.
- A class was abused as a namespace.

For that reason the constants now reside on the module global namespace,
and can be imported separately.

Documentation and demo code have been modified accordingly.