Improve and simplify logging.basicConfig()
If the module user does not want to use the provided logging facility,
probably a quite default logging experience is expected, and the root
logger will be used directly.

So, the new basicConfig() config is more similar to the usual behavior
of non-configured logging system, while at the same time having a better
format for timestamps, and avoiding surprises if a third party module
configures the root logger.