Fix logging fallback configuration format
If the included exception hook is used but logging has not been properly
configured using 'logger.config()', the format expected by the exception
hook and the 'error()' function is not the one set by 'basicConfig()'.

Even though the module facilities are intended to be used with a logger
explicitly set-up throught 'logger.config()', everything should work as
intended (maybe losing some power, but that is all) without it.

So the fallback configuration has been fixed so just the logging message
is printed, which is enough for the error reporting facilities to work.
This causes the loss of some logging context information, but should not
be a huge problem since this fallback will be used only in two possible
scenarios. The first one is a program using the module without doing the
proper logger configuration, and if logging context loss is a problem,
then the program should configure logging properly. The second one is an
error which happens much before the program could set up logging. Here,
the error should not happen EVER and has to be fixed, so again is not a
huge problem.