Add helper for OSError string formatting
Until now, a default formatting for OSError exceptions was provided, to
wit, function prettyprint_oserror(), but it not only did the formatting,
it also dumped the formatted string to logger.error(). This, in itself,
it not entirely desirable but also made impossible to further modify the
resulting string. It was convenient for some usage patterns, but it was
not a good idea.

As such, a new function called format_oserror() has been added and the
old prettyprint_oserror() has been removed.

Some system scripts have been modified to adapt to this new strategy.