Add py.typed to mark the module as type annotated
Some static type analyzers complain about a missing type stub for this
module, even though the module itself exposes typing annotations. So, in
order to mark the module as "type annotated", a "py.typed" file has to
be included in the installation.

Since this module as a flat structure, the "pyproject.toml" config file
has to be adapted so the package data (the "py.typed" file) is properly
installed alongside the module.